Introduce richer type for a set of decks
Previosusly a set of decks was just of type Deck[].  Deck type
filtering was done based on the type property of each deck.
The new type DecksCollection adds for explicitly defining deck
filters, possibly with some decks included in multiple categories.
An obvious example of this is the "all decks" filter.